% Auto-generated: do not edit by hand
\name{dragFrom}

\alias{dragFrom}

\title{DragFrom component}

\description{

}

\usage{
dragFrom(children=NULL, id=NULL, effectAllowed=NULL, name=NULL,
transferType=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{effectAllowed}{Character. }

\item{name}{Character. }

\item{transferType}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

