% Auto-generated: do not edit by hand
\name{flexLayout}

\alias{flexLayout}

\title{FlexLayout component}

\description{
Component description
}

\usage{
flexLayout(id=NULL)
}

\arguments{
\item{id}{Character. Unique ID to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

