% Auto-generated: do not edit by hand
\name{mantineProvider}

\alias{mantineProvider}

\title{MantineProvider component}

\description{
Mantine Provider
}

\usage{
mantineProvider(children=NULL, id=NULL, inherit=NULL, theme=NULL,
withCSSVariables=NULL, withGlobalStyles=NULL,
withNormalizeCSS=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Children}

\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{inherit}{Logical. Inherit from one level up MantineProvider}

\item{theme}{Named list. Theme}

\item{withCSSVariables}{Logical. With css variables}

\item{withGlobalStyles}{Logical. With global styles}

\item{withNormalizeCSS}{Logical. Normalize CSS}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

