% Auto-generated: do not edit by hand
\name{tab}

\alias{tab}

\title{Tab component}

\description{
Component description
}

\usage{
tab(children=NULL, id=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Children}

\item{id}{Character. Unique ID to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

