% Auto-generated: do not edit by hand
\name{flexLayout}

\alias{flexLayout}

\title{FlexLayout component}

\description{
Component description
}

\usage{
flexLayout(children=NULL, id=NULL, config=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Layout nodes}

\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{config}{Logical | numeric | character | named list | unnamed list. Layout config}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

